Create README.md
While i am having fun with this github as a beginer, i forked the master repo and added some other feature to the coming-soon-page. I acttually 
replaced the background image with a background video. Also, i noticed there is no favicon in the master, so i added it in the branch. I made some minor change to the countdown and the background 
overlay opacity. This is my second active day on github, so forgive my errors if there's any....i am a beginer.